restore is_erasable check to rmdir function
Otherwise the save_array doc example fails to write to a ZipStore

+++ zarr/storage.py
@@ -133,7 +133,7 @@ def rmdir(store: StoreLike, path: Path = None):
     this will be called, otherwise will fall back to implementation via the
     `Store` interface."""
     path = normalize_storage_path(path)
-    if hasattr(store, "rmdir"):
+    if hasattr(store, "rmdir") and store.is_erasable():  # type: ignore
         # pass through
         store.rmdir(path)  # type: ignore
     else:
